refactor(account): split service.go below 1000 lines
Move reconcile, import, and discovery helpers into sibling files.
No behavior change.

Related: #652
