+++ internal/account/service.go
@@ -3,7 +3,6 @@ package account
 import (
 	"context"
 	"database/sql"
-	"errors"
 	"fmt"
 	"net/url"
 	"slices"
@@ -18,10 +17,6 @@ import (
 )
 
 const defaultCalendarColor = "#7C3AED"
-
-// ErrSelectionStale means the account's imported calendar set changed after
-// discovery. Apply of the old checklist could then remove an unseen calendar.
-var ErrSelectionStale = errors.New("calendar selection is stale")
 
 type discoverFunc func(context.Context, Account, auth.Credential, func(auth.Credential) error) ([]caldav.RemoteCalendar, error)
 
@@ -280,442 +275,6 @@ func (s *Service) storeCredentialLocked(ctx context.Context, accountID int64, ex
 	return nil
 }
 
-// Discover retrieves a complete remote inventory and reconciles metadata for
-// already-imported calendars. Missing flags change only after the remote
-// discovery succeeds. Transient and partial failures then preserve local state.
-func (s *Service) Discover(ctx context.Context, accountID int64, store auth.CredentialStore) (Discovery, error) {
-	release, err := synclock.Account(ctx, s.db, accountID)
-	if err != nil {
-		return Discovery{}, fmt.Errorf("lock account discovery: %w", err)
-	}
-	defer release()
-	return s.discoverLocked(ctx, accountID, store)
-}
-
-// DiscoverWithCredential replaces an account's credential and runs a
-// complete discovery under the same lifecycle lock. A failed discovery restores
-// the previous credential. A reconnect with a typo then cannot break live sync.
-func (s *Service) DiscoverWithCredential(ctx context.Context, accountID int64, replacement auth.Credential, store auth.CredentialStore) (Discovery, error) {
-	release, err := synclock.Account(ctx, s.db, accountID)
-	if err != nil {
-		return Discovery{}, fmt.Errorf("lock account credential discovery: %w", err)
-	}
-	defer release()
-
-	configured, err := s.Get(ctx, accountID)
-	if err != nil {
-		return Discovery{}, fmt.Errorf("get account: %w", err)
-	}
-	fingerprint := configured.CredentialFingerprint()
-	previous, err := store.Get(accountID, fingerprint)
-	if err != nil {
-		return Discovery{}, fmt.Errorf("get previous account credentials: %w", err)
-	}
-	seedCredentialIdentity(&replacement, configured)
-	if replacement.RefreshToken == "" {
-		replacement.RefreshToken = previous.RefreshToken
-	}
-	if err := store.Set(replacement); err != nil {
-		return Discovery{}, fmt.Errorf("store replacement account credentials: %w", err)
-	}
-	discovery, err := s.discoverLocked(ctx, accountID, store)
-	if err == nil {
-		return discovery, nil
-	}
-	if restoreErr := store.Set(previous); restoreErr != nil {
-		return Discovery{}, fmt.Errorf("%w (restore previous account credentials: %w)", err, restoreErr)
-	}
-	return Discovery{}, err
-}
-
-// discoverLocked performs discovery and reconciliation while the caller holds
-// the account lifecycle lock.
-func (s *Service) discoverLocked(ctx context.Context, accountID int64, store auth.CredentialStore) (Discovery, error) {
-	account, err := s.Get(ctx, accountID)
-	if err != nil {
-		return Discovery{}, fmt.Errorf("get account: %w", err)
-	}
-	cred, err := store.Get(accountID, account.CredentialFingerprint())
-	if err != nil {
-		return Discovery{}, fmt.Errorf("get account credentials: %w", err)
-	}
-	found, err := s.discover(ctx, account, cred, func(updated auth.Credential) error {
-		updated.AccountID = accountID
-		updated.AccountFingerprint = account.CredentialFingerprint()
-		return store.Set(updated)
-	})
-	if err != nil {
-		return Discovery{}, fmt.Errorf("discover calendars: %w", err)
-	}
-
-	tx, err := s.db.BeginTx(ctx, nil)
-	if err != nil {
-		return Discovery{}, fmt.Errorf("begin discovery reconciliation: %w", err)
-	}
-	defer tx.Rollback()
-	qtx := s.q.WithTx(tx)
-	existingRows, err := qtx.ListCalendarsByAccount(ctx, &accountID)
-	if err != nil {
-		return Discovery{}, fmt.Errorf("list account calendars: %w", err)
-	}
-	existingByURL := make(map[string]storage.Calendar, len(existingRows))
-	for _, row := range existingRows {
-		existingByURL[remoteIdentityKey(storage.NullableToString(row.RemoteUrl), account.ServerURL)] = row
-	}
-	if err := qtx.MarkAccountCalendarsMissing(ctx, &accountID); err != nil {
-		return Discovery{}, fmt.Errorf("mark account calendars missing: %w", err)
-	}
-
-	calendars := make([]DiscoveredCalendar, 0, len(found))
-	seen := make(map[string]struct{}, len(found))
-	for _, remote := range found {
-		key := remoteIdentityKey(remote.Path, account.ServerURL)
-		if _, duplicate := seen[key]; duplicate {
-			continue
-		}
-		seen[key] = struct{}{}
-		remote.Name = remoteCalendarName(remote)
-		remote.Access = normalizedAccess(remote.Access)
-		remote.SupportedComponentSet = normalizedComponents(remote.SupportedComponentSet)
-		item := DiscoveredCalendar{RemoteCalendar: remote, Importable: supportsChroncal(remote.SupportedComponentSet)}
-		if local, ok := existingByURL[key]; ok {
-			item.Imported = true
-			item.CalendarID = local.ID
-			if err := applyDiscoveredCalendarMetadata(ctx, qtx, item, local.ID); err != nil {
-				return Discovery{}, fmt.Errorf("update discovered calendar %q: %w", remote.Name, err)
-			}
-		}
-		calendars = append(calendars, item)
-	}
-	for _, local := range existingRows {
-		path := storage.NullableToString(local.RemoteUrl)
-		if _, found := seen[remoteIdentityKey(path, account.ServerURL)]; found {
-			continue
-		}
-		name := strings.TrimSpace(local.RemoteName)
-		if name == "" {
-			name = local.Name
-		}
-		color := storage.NullableToString(local.RemoteColor)
-		if color == "" {
-			color = local.Color
-		}
-		components := normalizedComponents(strings.Split(local.RemoteComponents, ","))
-		calendars = append(calendars, DiscoveredCalendar{
-			RemoteCalendar: caldav.RemoteCalendar{
-				Path:                  path,
-				Name:                  name,
-				Description:           storage.NullableToString(local.Description),
-				Color:                 color,
-				Access:                normalizedAccess(caldav.CalendarAccess(local.RemoteAccess)),
-				SupportedComponentSet: components,
-			},
-			CalendarID: local.ID,
-			Imported:   true,
-			Importable: supportsChroncal(components),
-			Missing:    true,
-		})
-	}
-	if err := tx.Commit(); err != nil {
-		return Discovery{}, fmt.Errorf("commit discovery reconciliation: %w", err)
-	}
-	return Discovery{Account: account, Calendars: calendars}, nil
-}
-
-// Import creates local calendars for the selected paths. A repeat of an import
-// is idempotent. It returns the already-linked IDs instead of duplicate rows.
-func (s *Service) Import(ctx context.Context, discovery Discovery, selectedPaths []string) (ImportResult, error) {
-	if discovery.Account.ID == 0 {
-		return ImportResult{}, fmt.Errorf("discovery account is required")
-	}
-	release, err := synclock.Account(ctx, s.db, discovery.Account.ID)
-	if err != nil {
-		return ImportResult{}, fmt.Errorf("lock account import: %w", err)
-	}
-	defer release()
-
-	if _, err := s.q.GetAccount(ctx, discovery.Account.ID); err != nil {
-		return ImportResult{}, fmt.Errorf("get discovery account: %w", err)
-	}
-	byPath := make(map[string]DiscoveredCalendar, len(discovery.Calendars))
-	for _, item := range discovery.Calendars {
-		byPath[item.Path] = item
-	}
-
-	tx, err := s.db.BeginTx(ctx, nil)
-	if err != nil {
-		return ImportResult{}, fmt.Errorf("begin calendar import: %w", err)
-	}
-	defer tx.Rollback()
-	qtx := s.q.WithTx(tx)
-	existingRows, err := qtx.ListCalendarsByAccount(ctx, &discovery.Account.ID)
-	if err != nil {
-		return ImportResult{}, fmt.Errorf("list existing account calendars: %w", err)
-	}
-	existingByURL := make(map[string]int64, len(existingRows))
-	for _, row := range existingRows {
-		existingByURL[remoteIdentityKey(storage.NullableToString(row.RemoteUrl), discovery.Account.ServerURL)] = row.ID
-	}
-	// calendars.name is UNIQUE across the whole table, so a local display name
-	// must not collide with any existing calendar — local, linked to another
-	// account, or already created in this import batch. Seed the reserved set
-	// from every existing name and append a counter suffix on collision while
-	// keeping the pristine remote name in remote_name.
-	allCalendars, err := qtx.ListCalendars(ctx)
-	if err != nil {
-		return ImportResult{}, fmt.Errorf("list existing calendar names: %w", err)
-	}
-	taken := make(map[string]struct{}, len(allCalendars))
-	for _, row := range allCalendars {
-		taken[row.Name] = struct{}{}
-	}
-	unlinkedByURL := uniqueUnlinkedByRemoteIdentity(allCalendars, discovery.Account.ServerURL)
-
-	result := ImportResult{}
-	selected := make(map[string]struct{}, len(selectedPaths))
-	for _, path := range selectedPaths {
-		if _, duplicate := selected[path]; duplicate {
-			continue
-		}
-		selected[path] = struct{}{}
-		item, ok := byPath[path]
-		if !ok {
-			return ImportResult{}, fmt.Errorf("calendar %q was not part of this discovery", path)
-		}
-		if !item.Importable {
-			return ImportResult{}, fmt.Errorf("calendar %q has no supported event, todo, or journal components", item.Name)
-		}
-		key := remoteIdentityKey(path, discovery.Account.ServerURL)
-		if id, ok := existingByURL[key]; ok {
-			result.ExistingIDs = append(result.ExistingIDs, id)
-			continue
-		}
-		if id, ok := unlinkedByURL[key]; ok {
-			if err := relinkDiscoveredCalendar(ctx, qtx, discovery.Account, item, id); err != nil {
-				return ImportResult{}, fmt.Errorf("re-link calendar %q: %w", item.Name, err)
-			}
-			existingByURL[key] = id
-			delete(unlinkedByURL, key)
-			result.CreatedIDs = append(result.CreatedIDs, id)
-			continue
-		}
-
-		row, err := createDiscoveredCalendarRow(ctx, qtx, discovery.Account, item, taken)
-		if err != nil {
-			return ImportResult{}, fmt.Errorf("import calendar %q: %w", item.Name, err)
-		}
-		taken[row.Name] = struct{}{}
-		existingByURL[remoteIdentityKey(path, discovery.Account.ServerURL)] = row.ID
-		result.CreatedIDs = append(result.CreatedIDs, row.ID)
-	}
-	if err := tx.Commit(); err != nil {
-		return ImportResult{}, fmt.Errorf("commit calendar import: %w", err)
-	}
-	return result, nil
-}
-
-// ReconcileSelection atomically changes one account's local calendar set to
-// match the checked paths from a complete discovery. Remote collections are
-// never deleted. If no paths remain, the now-empty account and credential are
-// removed as part of the same operation.
-func (s *Service) ReconcileSelection(
-	ctx context.Context,
-	discovery Discovery,
-	params SelectionParams,
-	store auth.CredentialStore,
-) (SelectionResult, error) {
-	if discovery.Account.ID == 0 {
-		return SelectionResult{}, fmt.Errorf("discovery account is required")
-	}
-	release, err := synclock.Account(ctx, s.db, discovery.Account.ID)
-	if err != nil {
-		return SelectionResult{}, fmt.Errorf("lock account calendar selection: %w", err)
-	}
-	defer release()
-
-	configured, err := s.Get(ctx, discovery.Account.ID)
-	if err != nil {
-		return SelectionResult{}, fmt.Errorf("get discovery account: %w", err)
-	}
-	if configured.CredentialFingerprint() != discovery.Account.CredentialFingerprint() {
-		return SelectionResult{}, fmt.Errorf("%w: account connection changed", ErrSelectionStale)
-	}
-
-	discoveredByKey := make(map[string]DiscoveredCalendar, len(discovery.Calendars))
-	for _, item := range discovery.Calendars {
-		discoveredByKey[remoteIdentityKey(item.Path, discovery.Account.ServerURL)] = item
-	}
-	selected := make(map[string]struct{}, len(params.SelectedPaths))
-	selectedKeys := make([]string, 0, len(params.SelectedPaths))
-	for _, path := range params.SelectedPaths {
-		key := remoteIdentityKey(path, discovery.Account.ServerURL)
-		if _, duplicate := selected[key]; duplicate {
-			continue
-		}
-		item, ok := discoveredByKey[key]
-		if !ok {
-			return SelectionResult{}, fmt.Errorf("calendar %q was not part of this discovery", path)
-		}
-		if !item.Imported && (!item.Importable || item.Missing) {
-			return SelectionResult{}, fmt.Errorf("calendar %q cannot be added", item.Name)
-		}
-		selected[key] = struct{}{}
-		selectedKeys = append(selectedKeys, key)
-	}
-
-	removeAccount := len(selectedKeys) == 0
-	var prior auth.PriorCredential
-	if removeAccount {
-		if store == nil {
-			return SelectionResult{}, fmt.Errorf("credential store is required to remove an empty account")
-		}
-		prior, err = auth.CapturePriorCredential(store, discovery.Account.ID, discovery.Account.CredentialFingerprint())
-		if err != nil {
-			return SelectionResult{}, fmt.Errorf("read account credentials before removal: %w", err)
-		}
-	}
-
-	tx, err := s.db.BeginTx(ctx, nil)
-	if err != nil {
-		return SelectionResult{}, fmt.Errorf("begin account calendar reconciliation: %w", err)
-	}
-	defer tx.Rollback()
-	qtx := s.q.WithTx(tx)
-
-	existingRows, err := qtx.ListCalendarsByAccount(ctx, &discovery.Account.ID)
-	if err != nil {
-		return SelectionResult{}, fmt.Errorf("list account calendars: %w", err)
-	}
-	existingByKey := make(map[string]storage.Calendar, len(existingRows))
-	for _, row := range existingRows {
-		key := remoteIdentityKey(storage.NullableToString(row.RemoteUrl), discovery.Account.ServerURL)
-		item, ok := discoveredByKey[key]
-		if !ok || !item.Imported || item.CalendarID != row.ID {
-			return SelectionResult{}, fmt.Errorf("%w: imported calendars changed", ErrSelectionStale)
-		}
-		existingByKey[key] = row
-	}
-	for key, item := range discoveredByKey {
-		if !item.Imported {
-			continue
-		}
-		row, ok := existingByKey[key]
-		if !ok || row.ID != item.CalendarID {
-			return SelectionResult{}, fmt.Errorf("%w: imported calendars changed", ErrSelectionStale)
-		}
-	}
-
-	removedRows := make([]storage.Calendar, 0, len(existingRows))
-	removedIDs := make(map[int64]struct{}, len(existingRows))
-	finalByKey := make(map[string]int64, len(selectedKeys))
-	for key, row := range existingByKey {
-		if _, keep := selected[key]; keep {
-			finalByKey[key] = row.ID
-			continue
-		}
-		removedRows = append(removedRows, row)
-		removedIDs[row.ID] = struct{}{}
-	}
-	addedCount := 0
-	for _, key := range selectedKeys {
-		if _, exists := existingByKey[key]; !exists {
-			addedCount++
-		}
-	}
-
-	allCalendars, err := qtx.ListCalendars(ctx)
-	if err != nil {
-		return SelectionResult{}, fmt.Errorf("list existing calendars: %w", err)
-	}
-	if len(allCalendars)-len(removedRows)+addedCount < 1 {
-		return SelectionResult{}, calendar.ErrLastCalendar
-	}
-
-	var removedDefault bool
-	for _, row := range removedRows {
-		if row.IsDefault == 1 {
-			removedDefault = true
-			break
-		}
-	}
-	if removedDefault && params.NewDefaultID == 0 && strings.TrimSpace(params.NewDefaultPath) == "" {
-		return SelectionResult{}, calendar.ErrDefaultCalendarRequiresPromotion
-	}
-	if removedDefault && params.NewDefaultID != 0 && strings.TrimSpace(params.NewDefaultPath) != "" {
-		return SelectionResult{}, calendar.ErrInvalidPromotionTarget
-	}
-
-	replacementID := params.NewDefaultID
-
-	taken := make(map[string]struct{}, len(allCalendars))
-	for _, row := range allCalendars {
-		if _, removed := removedIDs[row.ID]; !removed {
-			taken[row.Name] = struct{}{}
-		}
-	}
-	unlinkedByURL := uniqueUnlinkedByRemoteIdentity(allCalendars, discovery.Account.ServerURL)
-
-	result := SelectionResult{RemovedIDs: make([]int64, 0, len(removedRows))}
-	for _, row := range removedRows {
-		if err := qtx.DeleteCalendar(ctx, row.ID); err != nil {
-			return SelectionResult{}, fmt.Errorf("remove calendar %q: %w", row.Name, err)
-		}
-		result.RemovedIDs = append(result.RemovedIDs, row.ID)
-	}
-
-	for _, key := range selectedKeys {
-		if _, exists := finalByKey[key]; exists {
-			continue
-		}
-		item := discoveredByKey[key]
-		if id, ok := unlinkedByURL[key]; ok {
-			if err := relinkDiscoveredCalendar(ctx, qtx, discovery.Account, item, id); err != nil {
-				return SelectionResult{}, fmt.Errorf("re-link calendar %q: %w", item.Name, err)
-			}
-			delete(unlinkedByURL, key)
-			finalByKey[key] = id
-			result.CreatedIDs = append(result.CreatedIDs, id)
-			continue
-		}
-		row, err := createDiscoveredCalendarRow(ctx, qtx, discovery.Account, item, taken)
-		if err != nil {
-			return SelectionResult{}, fmt.Errorf("add calendar %q: %w", item.Name, err)
-		}
-		taken[row.Name] = struct{}{}
-		finalByKey[key] = row.ID
-		result.CreatedIDs = append(result.CreatedIDs, row.ID)
-	}
-
-	if removedDefault {
-		if path := strings.TrimSpace(params.NewDefaultPath); path != "" {
-			key := remoteIdentityKey(path, discovery.Account.ServerURL)
-			var ok bool
-			replacementID, ok = finalByKey[key]
-			if !ok {
-				return SelectionResult{}, calendar.ErrInvalidPromotionTarget
-			}
-		}
-		if err := calendar.PromoteDefault(ctx, qtx, removedIDs, replacementID); err != nil {
-			return SelectionResult{}, err
-		}
-	}
-
-	if removeAccount {
-		if err := qtx.DeleteAccount(ctx, discovery.Account.ID); err != nil {
-			return SelectionResult{}, fmt.Errorf("remove empty account: %w", err)
-		}
-		if err := store.Delete(discovery.Account.ID); err != nil {
-			return SelectionResult{}, prior.Restore(store, discovery.Account.ID, false, "delete empty account credentials", err)
-		}
-		result.AccountRemoved = true
-	}
-
-	if err := auth.CommitWithCredentialCompensation(tx, store, discovery.Account.ID, prior, false, "commit account calendar reconciliation"); err != nil {
-		return SelectionResult{}, err
-	}
-	return result, nil
-}
-
 // RemoveWithCalendars deletes an account and every local calendar attached to
 // it. It never contacts the remote server. Delete has a different contract:
 // that method preserves the calendars as disconnected local rows.
@@ -865,17 +424,6 @@ func (s *Service) Delete(ctx context.Context, accountID int64, store auth.Creden
 	return nil
 }
 
-func discoverRemoteCalendars(ctx context.Context, account Account, cred auth.Credential, persist func(auth.Credential) error) ([]caldav.RemoteCalendar, error) {
-	if caldav.IsGoogleCalendarEndpoint(account.ServerURL) && cred.AccessToken != "" {
-		return caldav.DiscoverGoogleCalendars(ctx, cred, persist)
-	}
-	client, err := caldav.NewClientFromCredential(account.ServerURL, cred, persist)
-	if err != nil {
-		return nil, err
-	}
-	return client.DiscoverCalendars(ctx)
-}
-
 func validateServerURL(raw string, allowInsecure bool) (string, error) {
 	raw = strings.TrimSpace(raw)
 	parsed, err := url.Parse(raw)
@@ -923,121 +471,6 @@ func normalizedComponents(components []string) []string {
 	}
 	slices.Sort(out)
 	return out
-}
-
-// uniqueUnlinkedByRemoteIdentity maps a remote identity to one unlinked local
-// calendar. Duplicate keys are dropped so Import and ReconcileSelection create
-// a fresh row. They must not pick an arbitrary snapshot.
-func uniqueUnlinkedByRemoteIdentity(rows []storage.Calendar, serverURL string) map[string]int64 {
-	byKey := make(map[string]int64)
-	ambiguous := make(map[string]struct{})
-	for _, row := range rows {
-		if row.AccountID != nil {
-			continue
-		}
-		key := remoteIdentityKey(storage.NullableToString(row.RemoteUrl), serverURL)
-		if key == "" {
-			continue
-		}
-		if _, taken := byKey[key]; taken {
-			delete(byKey, key)
-			ambiguous[key] = struct{}{}
-			continue
-		}
-		if _, skip := ambiguous[key]; skip {
-			continue
-		}
-		byKey[key] = row.ID
-	}
-	return byKey
-}
-
-// applyDiscoveredCalendarMetadata writes live discovery color, name, access,
-// and components onto a local calendar. AdoptCalendarRemoteName keeps a
-// user-chosen display name. UpdateCalendarDiscovery is the one write of the
-// remote_* mirrors and the color CASE rules.
-func applyDiscoveredCalendarMetadata(ctx context.Context, qtx *storage.Queries, item DiscoveredCalendar, calendarID int64) error {
-	remoteName := remoteCalendarName(item.RemoteCalendar)
-	if err := qtx.AdoptCalendarRemoteName(ctx, storage.AdoptCalendarRemoteNameParams{
-		Name: remoteName,
-		ID:   calendarID,
-	}); err != nil {
-		return err
-	}
-	_, err := qtx.UpdateCalendarDiscovery(ctx, storage.UpdateCalendarDiscoveryParams{
-		RemoteName:       remoteName,
-		RemoteColor:      item.Color,
-		RemoteAccess:     string(normalizedAccess(item.Access)),
-		RemoteComponents: strings.Join(normalizedComponents(item.SupportedComponentSet), ","),
-		ID:               calendarID,
-	})
-	return err
-}
-
-// relinkDiscoveredCalendar attaches one unlinked local calendar to the account.
-// It writes the live collection onto the row through applyDiscoveredCalendarMetadata.
-func relinkDiscoveredCalendar(ctx context.Context, qtx *storage.Queries, acct Account, item DiscoveredCalendar, calendarID int64) error {
-	accountID := acct.ID
-	if err := qtx.LinkCalendarToAccount(ctx, storage.LinkCalendarToAccountParams{
-		ID:        calendarID,
-		AccountID: &accountID,
-		RemoteUrl: storage.StringToNullable(item.Path),
-	}); err != nil {
-		return err
-	}
-	if err := applyDiscoveredCalendarMetadata(ctx, qtx, item, calendarID); err != nil {
-		return err
-	}
-	return qtx.UpdateCalendarOwnerEmail(ctx, storage.UpdateCalendarOwnerEmailParams{
-		OwnerEmail: acct.Username,
-		ID:         calendarID,
-	})
-}
-
-// createDiscoveredCalendarRow creates the local calendar row for a discovered
-// remote collection. It reserves a local name from taken that does not collide.
-// The caller adds the returned row's name to taken when it creates in a loop.
-// It is the single DiscoveredCalendar-to-row mapping shared by Import,
-// ReconcileSelection, and calendar migration. Remote-metadata columns then stay
-// in lockstep no matter which flow creates the row.
-func createDiscoveredCalendarRow(
-	ctx context.Context,
-	qtx *storage.Queries,
-	acct Account,
-	item DiscoveredCalendar,
-	taken map[string]struct{},
-) (storage.Calendar, error) {
-	remoteName := remoteCalendarName(item.RemoteCalendar)
-	color := item.Color
-	if color == "" {
-		color = defaultCalendarColor
-	}
-	accountID := acct.ID
-	return qtx.CreateDiscoveredCalendar(ctx, storage.CreateDiscoveredCalendarParams{
-		Name:             uniqueLocalName(remoteName, taken),
-		Color:            color,
-		Description:      storage.StringToNullable(item.Description),
-		AccountID:        &accountID,
-		RemoteUrl:        storage.StringToNullable(item.Path),
-		RemoteColor:      storage.StringToNullable(item.Color),
-		RemoteName:       remoteName,
-		RemoteAccess:     string(normalizedAccess(item.Access)),
-		RemoteComponents: strings.Join(normalizedComponents(item.SupportedComponentSet), ","),
-		OwnerEmail:       acct.Username,
-	})
-}
-
-func supportsChroncal(components []string) bool {
-	if len(components) == 0 {
-		return true
-	}
-	for _, component := range components {
-		switch component {
-		case "VEVENT", "VTODO", "VJOURNAL":
-			return true
-		}
-	}
-	return false
 }
 
 func remoteCalendarName(remote caldav.RemoteCalendar) string {
